license/requirements.txt
aniso8601==7.0.0
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2022.9.14
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.4
django-reset-migrations==0.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.2.0
promise==2.3
PyJWT==2.5.0
pyparsing==3.0.9
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.12
